pyproject.toml
[project]
name = "ufo-watcher"
version = "0.1.0"
description = "Desktop UFO that flies when a URL changes"
requires-python = ">=3.9"
dependencies = [
    "pyobjc-framework-Cocoa>=10.0",
    "beautifulsoup4>=4.12",
]

[tool.uv]
dev-dependencies = []
